--- layers/Schema/packages/migrate-taxonomies-wp/.gitattributes
@@ -1,16 +1,0 @@
-# Path-based git attributes
-# https://www.kernel.org/pub/software/scm/git/docs/gitattributes.html
-
-# Ignore all test and documentation with "export-ignore".
-/.editorconfig               export-ignore
-/.gitattributes              export-ignore
-/.gitignore                  export-ignore
-/.scrutinizer.yml            export-ignore
-/.styleci.yml                export-ignore
-/.travis.yml                 export-ignore
-/PULL_REQUEST_TEMPLATE.md    export-ignore
-/ISSUE_TEMPLATE.md           export-ignore
-/phpcs.xml.dist              export-ignore
-/phpunit.xml.dist            export-ignore
-/tests                       export-ignore
-/docs                        export-ignore
