--- layers/Schema/packages/migrate-taxonomies-wp/.gitignore
@@ -1,6 +1,0 @@
-/build/
-composer.lock
-/vendor/
-phpcs.xml
-phpunit.xml
-.phpunit.result.cache
