--- layers/Schema/packages/migrate-taxonomies-wp/composer.json
@@ -1,28 +1,0 @@
-{
-    "name": "pop-schema/migrate-taxonomies-wp",
-    "type": "library",
-    "description": "(Temporary package) Migrate code to package https://github.com/pop-schema/taxonomies-wp. Target package description: Implementation for WordPress of contracts from package \"Taxonomies\"",
-    "keywords": ["PoP", "taxonomies-wp"],
-    "homepage": "https://github.com/pop-schema/migrate-taxonomies-wp",
-    "license": "GPL-2.0-or-later",
-    "authors": [
-        {
-            "name": "Leonardo Losoviz",
-            "email": "leo@getpop.org",
-            "homepage": "https://getpop.org"
-        }
-    ],
-    "minimum-stability": "dev",
-    "prefer-stable": true,
-    "require": {
-        "php": "^8.0"
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "0.8-dev"
-        }
-    },
-    "config": {
-        "sort-packages": true
-    }
-}
