--- layers/Schema/packages/migrate-taxonomies-wp/migrate/contractimplementations/cms-apis/load.php
@@ -1,3 +1,0 @@
-<?php
-
-require_once 'cms-functionapi.php';
