--- layers/Schema/packages/migrate-taxonomies-wp/migrate/contractimplementations/load.php
@@ -1,3 +1,0 @@
-<?php
-
-require_once 'cms-apis/load.php';
