--- layers/Schema/packages/migrate-taxonomies-wp/migrate/initialization.php
@@ -1,15 +1,0 @@
-<?php
-namespace PoPSchema\Taxonomies\WP;
-
-class Initialization
-{
-    public function initialize()
-    {
-        load_plugin_textdomain('pop-taxonomies-wp', false, dirname(plugin_basename(__FILE__)).'/languages');
-
-        /**
-         * Load the Contract Implementations
-         */
-        include_once 'contractimplementations/load.php';
-    }
-}
