+++ layers/Schema/packages/taxonomies-wp/composer.json
@@ -17,8 +17,7 @@
     "require": {
         "php": "^8.0",
         "pop-schema/taxonomies": "^0.8",
-        "pop-schema/customposts-wp": "^0.8",
-        "pop-schema/migrate-taxonomies-wp": "^0.8"
+        "pop-schema/customposts-wp": "^0.8"
     },
     "require-dev": {
         "phpunit/phpunit": ">=9.3",
