+++ webservers/graphql-api-for-wp/.lando.yml
@@ -193,8 +193,6 @@ services:
         - >-
           ../../layers/Schema/packages/migrate-taxonomies:/app/wordpress/wp-content/plugins/graphql-api/vendor/pop-schema/migrate-taxonomies
         - >-
-          ../../layers/Schema/packages/migrate-taxonomies-wp:/app/wordpress/wp-content/plugins/graphql-api/vendor/pop-schema/migrate-taxonomies-wp
-        - >-
           ../../layers/Schema/packages/migrate-taxonomymeta:/app/wordpress/wp-content/plugins/graphql-api/vendor/pop-schema/migrate-taxonomymeta
         - >-
           ../../layers/Schema/packages/migrate-taxonomymeta-wp:/app/wordpress/wp-content/plugins/graphql-api/vendor/pop-schema/migrate-taxonomymeta-wp
