+++ webservers/graphql-by-pop/.lando.yml
@@ -182,8 +182,6 @@ services:
         - >-
           ../../layers/Schema/packages/migrate-taxonomies:/app/vendor/pop-schema/migrate-taxonomies
         - >-
-          ../../layers/Schema/packages/migrate-taxonomies-wp:/app/vendor/pop-schema/migrate-taxonomies-wp
-        - >-
           ../../layers/Schema/packages/migrate-taxonomymeta:/app/vendor/pop-schema/migrate-taxonomymeta
         - >-
           ../../layers/Schema/packages/migrate-taxonomymeta-wp:/app/vendor/pop-schema/migrate-taxonomymeta-wp
